moved auth fields to location rather than server blocks
